fix word size detection
Due to the order in which jconfig.h was included in
jinclude.h, __WORDSIZE was not yet defined, so that
SIZEOF_SIZE_T was always defined as 4 even on 64bit architectures.
Also jinclude.h and jconfig.h had no include guard and where
included more than once which lead to redefinitions of the
SIZEOF_SIZE_T macro.

This patch
 * makes SIZEOF_SIZE_T defined as __SIZEOF_SIZE_T__, which is predefined
   by the compiler,
 * makes the preprocessor emit an error if __SIZEOF_SIZE_T__ is not
   predefined, and
 * adds an include guard to jconfig.h.

Change-Id: Icb029a6faafdc847e2b735bd9adb465868f1b17d