Refactor tjunittest.c to provide test interface
tjunittest.c contains unit tests for the TurboJPEG API (which in turn
test the underlying libjpeg-turbo code.) These tests were designed to
run as a standalone binary.

This commit refactors tjunittest.c to provide an interface to run all
unit tests individually. This is the first step towards running all
tests through the gtest framework.

The following files are also reinstated from the upstream
libjpeg-turbo project as the tests in tjunittest.c rely on an MD5
checksum:
 * md5/md5.h
 * md5/md5.c
 * md5/md5hl.c

Bug: 993876
Change-Id: I6380a8f0ca5db02b40a1235e379cb1e5134becca