Refactor tjbench.c to provide test interface
tjbench.c contains a program to benchmark encoding and decoding of
JPEG images. As such, the code in tjbench.c was designed to run as
a stanalone binary. As well as benchmarking code, tjbench is used
to produce image data for unit test input.

This commit refactors tjbench.c to provide an interface to run the
tjbench code programmatically. A gtest wrapper containing unit tests
will be introduced in a subsequent commit.

All other global symbols in tjbench.c are marked as 'static' to
prevent symbol collision.

Bug: 993876
Change-Id: Ib8469d4ee8adee466f0b9d8a9dd696bd5ef7c205