Add input JPEG images for djpeg and jpegtran tests
Upstream unit tests use the output of cjpeg tests as input for djpeg
and jpegtran tests. To avoid the coupling of compression and
decompression unit tests, this commit adds the input JPEG images for
the djpeg and jpegtran unit tests to the testimages directory.

These images are generated from the original images in the testimages
directory and are distributed under the same license.

Bug: 993876
Change-Id: I1c08432f8891f71d081df284ff83bbc021d8cb72