Disable Neon Huffman encoding for WoA/Clang-cl
The Neon SIMD path for Huffman encoding (baseline, not progressive)
does not work properly when compiled for Windows on Arm (WoA) using
Clang-cl.

For now - as the cause of the problem is not trivial to diagnose or
fix - disable the Neon Huffman encoding path when compiling for WoA
using Clang-cl.

Bug: 1160249
Change-Id: I6381cbe8cabc218ab6541f2b64fa07b9a83fcd12